Overhaul of design to include changes to tenting and errors made in measurements
1) added articulated legs for tenting. legs will use a lock washer such as a nordlock or internal
   shakeproof washer to prevent rotation once tightened.
2) fixes to hole placement in top plate
3) reduced spacing between top plate and PCB assembly
4) Added full assebly including base plate

I need to go through and tidy up the parameters used in the spreadsheets and possibly refactor them to be clearer.
The naming of lots of parameters is less than ideal

Some elements of the design are still subject to topological naming issues.

Changes to top plate have habit of rippling through design as this acts as master sketch
with subshape binder. Could consider moving to separate top plate body and maintain existing
body as reference.